feat:move `AlertHistory` into `AlertTable`
