+++ keep-ui/app/alerts/alert-table.tsx
@@ -47,6 +47,7 @@ import AlertColumnsSelect, {
 import AlertTableCheckbox from "./alert-table-checkbox";
 import AlertPagination from "./alert-pagination";
 import { KeyedMutator } from "swr";
+import { AlertHistory } from "./alert-history";
 
 const getAlertLastReceieved = (lastRecievedFromAlert: Date) => {
   let lastReceived = "unknown";
@@ -81,7 +82,6 @@ interface Props {
   ) => void;
   users?: User[];
   currentUser: NextUser;
-  openModal?: (alert: AlertDto) => void;
   presetName?: string;
   rowSelection?: RowSelectionState;
   setRowSelection?: OnChangeFn<RowSelectionState>;
@@ -99,12 +99,23 @@ export function AlertTable({
   setAssignee,
   users = [],
   currentUser,
-  openModal,
   presetName,
   rowSelection,
   setRowSelection,
 }: Props) {
   const router = useRouter();
+  const [isOpen, setIsOpen] = useState(false);
+  const [selectedAlertHistory, setSelectedAlertHistory] = useState<AlertDto[]>(
+    []
+  );
+
+  const openModal = (alert: AlertDto): any => {
+    setSelectedAlertHistory(groupedByAlerts[(alert as any)[groupBy!]]);
+    setIsOpen(true);
+  };
+
+  const enabledRowSelection =
+    presetName === "Deleted" || isOpen ? undefined : rowSelection;
 
   const handleWorkflowClick = (workflows: Workflow[]) => {
     if (workflows.length === 1) {
@@ -114,7 +125,7 @@ export function AlertTable({
     }
   };
 
-  const checkboxColumn = rowSelection
+  const checkboxColumn = enabledRowSelection
     ? [
         columnHelper.display({
           id: "checkbox",
@@ -150,7 +161,7 @@ export function AlertTable({
               canOpenHistory={
                 !groupedByAlerts![(context.row.original as any)[groupBy!]]
               }
-              openHistory={() => openModal!(context.row.original)}
+              openHistory={() => openModal(context.row.original)}
               provider={providers.find(
                 (p) => p.type === context.row.original.source![0]
               )}
@@ -322,7 +333,7 @@ export function AlertTable({
     state: {
       columnVisibility,
       columnOrder,
-      rowSelection,
+      rowSelection: enabledRowSelection,
     },
     initialState: {
       pagination: { pageSize: 10 },
@@ -377,6 +388,13 @@ export function AlertTable({
         <AlertsTableBody table={table} showSkeleton={isAsyncLoading} />
       </Table>
       <AlertPagination table={table} mutate={mutate} />
+      <AlertHistory
+        isOpen={isOpen}
+        closeModal={() => setIsOpen(false)}
+        data={selectedAlertHistory}
+        users={users}
+        currentUser={currentUser}
+      />
     </>
   );
 }
